// convex/_generated/api.d.ts
/* eslint-disable */
/**
 * Generated `api` utility.
 *
 * THIS CODE IS AUTOMATICALLY GENERATED.
 *
 * To regenerate, run `npx convex dev`.
 * @module
 */

import type {
  ApiFromModules,
  FilterApi,
  FunctionReference,
} from "convex/server";
import type * as Brand from "../Brand.js";
import type * as Cart from "../Cart.js";
import type * as Categories from "../Categories.js";
import type * as Product from "../Product.js";
import type * as Recipe from "../Recipe.js";
import type * as Storage from "../Storage.js";
import type * as users from "../users.js";

/**
 * A utility for referencing Convex functions in your app's API.
 *
 * Usage:
 * ```js
 * const myFunctionReference = api.myModule.myFunction;
 * ```
 */
declare const fullApi: ApiFromModules<{
  Brand: typeof Brand;
  Cart: typeof Cart;
  Categories: typeof Categories;
  Product: typeof Product;
  Recipe: typeof Recipe;
  Storage: typeof Storage;
  users: typeof users;
}>;
export declare const api: FilterApi<
  typeof fullApi,
  FunctionReference<any, "public">
>;
export declare const internal: FilterApi<
  typeof fullApi,
  FunctionReference<any, "internal">
>;